Use Symbolics.jl
